+++ tests/test_sdclocation.py
@@ -1,11 +1,12 @@
 import unittest
 
 from sdc11073.location import SdcLocation
-
+from sdc11073.wsdiscovery.service import Service
+from sdc11073.xml_types.wsd_types import ScopesType
 
 class TestSdcLocation(unittest.TestCase):
     scheme = SdcLocation.scheme  # 'sdc.ctxt.loc'
-    default_root = SdcLocation.location_detail_root  # 'sdc.ctxt.loc.detail'
+    default_root = 'sdc.ctxt.loc.detail'
     scope_prefix = scheme + ':/' + default_root  # sdc.ctxt.loc:/sdc.ctxt.loc.detail'
 
     def test_scopeString(self):
@@ -16,7 +17,7 @@ class TestSdcLocation(unittest.TestCase):
         self.assertEqual(loc.poc, 'CU1')
         self.assertEqual(loc.bed, 'BedA500')
         self.assertEqual(loc.rm, None)
-        self.assertEqual(loc.bld, None)
+        self.assertEqual(loc.bldng, None)
         self.assertEqual(loc.flr, None)
         self.assertEqual(loc.scope_string, expected_scope_string)
 
@@ -28,19 +29,19 @@ class TestSdcLocation(unittest.TestCase):
         self.assertEqual(loc.poc, None)
         self.assertEqual(loc.bed, 'BedA500')
         self.assertEqual(loc.rm, None)
-        self.assertEqual(loc.bld, None)
+        self.assertEqual(loc.bldng, None)
         self.assertEqual(loc.flr, None)
         self.assertEqual(loc.scope_string, expected_scope_string)
 
         # this is an unusual scope with all parameters and spaces in them
-        loc = SdcLocation(fac='HOSP 1', poc='CU 1', bed='Bed A500', flr='flr 1', rm='rM 1', bld='abc 1',
+        loc = SdcLocation(fac='HOSP 1', poc='CU 1', bed='Bed A500', flr='flr 1', rm='rM 1', bldng='abc 1',
                           root='some where')
         self.assertEqual(loc.root, 'some where')
         self.assertEqual(loc.fac, 'HOSP 1')
         self.assertEqual(loc.poc, 'CU 1')
         self.assertEqual(loc.bed, 'Bed A500')
         self.assertEqual(loc.rm, 'rM 1')
-        self.assertEqual(loc.bld, 'abc 1')
+        self.assertEqual(loc.bldng, 'abc 1')
         self.assertEqual(loc.flr, 'flr 1')
 
         self.assertEqual(loc, SdcLocation.from_scope_string(loc.scope_string))
@@ -53,7 +54,7 @@ class TestSdcLocation(unittest.TestCase):
         self.assertEqual(loc.poc, 'CU1')
         self.assertEqual(loc.bed, 'BedA500')
         self.assertEqual(loc.rm, None)
-        self.assertEqual(loc.bld, None)
+        self.assertEqual(loc.bldng, None)
         self.assertEqual(loc.flr, None)
         self.assertEqual(loc.scope_string, scope_string_sdc)
 
@@ -65,7 +66,7 @@ class TestSdcLocation(unittest.TestCase):
         self.assertEqual(loc.poc, 'CU 1')
         self.assertEqual(loc.bed, 'Bed A500')
         self.assertEqual(loc.rm, 'rM 1')
-        self.assertEqual(loc.bld, 'abc 1')
+        self.assertEqual(loc.bldng, 'abc 1')
         self.assertEqual(loc.flr, 'flr 1')
 
         # if we can create another identical  DraegerLocation from loc, then scopeString also seems okay.
@@ -81,14 +82,14 @@ class TestSdcLocation(unittest.TestCase):
             self.assertEqual(loc.poc, 'CU 1')
             self.assertEqual(loc.bed, 'Bed A500')
             self.assertEqual(loc.rm, 'rM 1')
-            self.assertEqual(loc.bld, 'abc 1')
+            self.assertEqual(loc.bldng, 'abc 1')
             self.assertEqual(loc.flr, 'flr 1')
 
     def test_equal(self):
         loc1 = SdcLocation(bed='BedA500', root='myroot')
         loc2 = SdcLocation(bed='BedA500', root='myroot')
         self.assertEqual(loc1, loc2)
-        for attrName in ('root', 'fac', 'bld', 'poc', 'flr', 'rm', 'bed'):
+        for attrName in ('root', 'fac', 'bldng', 'poc', 'flr', 'rm', 'bed'):
             print('different {} expected'.format(attrName))
             setattr(loc1, attrName, 'x')
             setattr(loc2, attrName, 'y')
@@ -99,9 +100,9 @@ class TestSdcLocation(unittest.TestCase):
 
     def test_contains(self):
         whole_world = SdcLocation()
-        my_bed = SdcLocation(fac='fac1', poc='poc1', bed='bed1', bld='bld1', flr='flr1', rm='rm1')
-        my_bld = SdcLocation(fac='fac1', poc='poc1', bld='bld1')
-        other_bld = SdcLocation(fac='fac1', poc='poc1', bld='bld2')
+        my_bed = SdcLocation(fac='fac1', poc='poc1', bed='bed1', bldng='bld1', flr='flr1', rm='rm1')
+        my_bld = SdcLocation(fac='fac1', poc='poc1', bldng='bld1')
+        other_bld = SdcLocation(fac='fac1', poc='poc1', bldng='bld2')
         any_flr1 = SdcLocation(flr='flr1')  # any location that has flr1 will match
         self.assertTrue(my_bed in whole_world)
         self.assertFalse(whole_world in my_bed)
@@ -116,17 +117,18 @@ class TestSdcLocation(unittest.TestCase):
         self.assertFalse(my_bed in other_bld)
 
         # non-default root
-        my_bed = SdcLocation(fac='fac1', poc='poc1', bed='bed1', bld='bld1', flr='flr1', rm='rm1', root='myroot')
+        my_bed = SdcLocation(fac='fac1', poc='poc1', bed='bed1', bldng='bld1', flr='flr1', rm='rm1', root='myroot')
         self.assertTrue(
-            my_bed in SdcLocation(fac='fac1', poc='poc1', bed='bed1', bld='bld1', flr='flr1', rm='rm1', root='myroot'))
+            my_bed in SdcLocation(fac='fac1', poc='poc1', bed='bed1', bldng='bld1', flr='flr1', rm='rm1', root='myroot'))
         self.assertTrue(my_bed in SdcLocation(fac='fac1', root='myroot'))
         self.assertFalse(my_bed in SdcLocation(fac='fac2', root='myroot'))
         self.assertFalse(my_bed in SdcLocation(fac='fac1'))
 
-
-def suite():
-    return unittest.TestLoader().loadTestsFromTestCase(TestSdcLocation)
-
-
-if __name__ == '__main__':
-    unittest.TextTestRunner(verbosity=2).run(suite())
+    def test_filter_services_inside(self):
+        my_loc = SdcLocation(fac='fac1', poc='poc1', bed='bed1', bldng='bld1', flr='flr1', rm='rm1', root='myroot')
+        other_loc = SdcLocation(fac='fac2', poc='poc1', bed='bed1', bldng='bld1', flr='flr1', rm='rm1', root='myroot')
+        service1 = Service(types=None, scopes=ScopesType(my_loc.scope_string), epr='a', x_addrs=None, instance_id='42')
+        service2 = Service(types=None, scopes=ScopesType(other_loc.scope_string), epr='b', x_addrs=None, instance_id='42')
+        service3 = Service(types=None, scopes=None, epr='b', x_addrs=None, instance_id='42')
+        matches = my_loc.filter_services_inside((service1, service2, service3))
+        self.assertEqual(len(matches), 1)