+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   sdc11073.mdib.clientmdib.ClientMdibContainer becomes sdc11073.mdib.consumermdib.ConsumerMdib, etc.
 - reduced max_subscription_duration of provider to 15 seconds
 - renew can be performed in fixed intervals (as before) or depending on Expires time of subscription.
+- SdcLocation class reworked; use 'bldng' instead of 'bld', which better matches the standard. 
 
 ## [2.0.0a5] - 2023-06-27